add integration tests
This test verifies that the output from this system is usable by the `go
get` tool. It operates by proxying `go get` HTTP requests to a locally
running proxy, then checking the URL the VCS cloned from.

This isn't a hermetic test: it requires internet access to github.com
and bitbucket.org, which I'm not sure how to reasonably fake and provide
the same level of confidence. It checks PATH for locally installed
copies of go, git, and hg, but those could be bundled in a test
environment.

I bumped up the version of Go required/tested against to Go 1.8 in order
to use subtests.  Go 1.8 is now GA on App Engine, so this should not be
a problem.