Update to use app engine runtime go112 (#34)
* app.yaml: update to go112

Make the required changes to use go112 due to the deprecation of Go 1.9.

Migration doc:
- https://cloud.google.com/appengine/docs/standard/go111/go-differences

References:
- https://github.com/GoogleCloudPlatform/govanityurls/issues/29#issuecomment-506538041
- https://github.com/etcd-io/maintainers/issues/17

* go.mod: initial commit

Migrate this project to use go.mod and update the README.

This is part of the recommended steps to runtime: go111 and beyond.
https://cloud.google.com/appengine/docs/standard/go111/go-differences

* vendor: add dependent packages

Add all dependencies for the project via vendoring for offline builds.

* travis.yml: drop go 1.6 for 1.12

Go 1.6 is too old to have a number of builtin packages for the latest
appengine package. Remove and update to match app engine version: 1.12.

* appengine.go: remove unused file

From the go111 migration doc:
"The appengine build tag is deprecated and will not be used when
building your app for deployment.  Ensure your code still functions
correctly without it being set."

https://cloud.google.com/appengine/docs/standard/go111/go-differences

* main: respect PORT environment variable

The go111 runtime should use the PORT variable to decide the listening
http port.

"PORT The port that receives HTTP requests."
https://cloud.google.com/appengine/docs/standard/go111/runtime